+++ packages/client/hmi-client/src/components/workflow/tera-simulate-ensemble-ciemss.vue
@@ -187,7 +187,8 @@ import { ref, computed, watch } from 'vue';
 import { getRunResultCiemss } from '@/services/models/simulation-service';
 import { getModelConfigurationById } from '@/services/model-configurations';
 import { WorkflowNode } from '@/types/workflow';
-import { workflowEventBus, saveDataset } from '@/services/workflow';
+import { workflowEventBus } from '@/services/workflow';
+import { saveDataset } from '@/services/dataset';
 import Button from 'primevue/button';
 import AccordionTab from 'primevue/accordiontab';
 import Accordion from 'primevue/accordion';
